Final fix for RGB decay and robust split-half rendering
- Fix decay stall bug by using uint32_t for diff and ticks, preventing unsigned underflow and truncation.
- Implement robust split-aware rendering using is_keyboard_master() to handle master/slave range differences correctly.
- Ensure Master calculates colors for all 58 LEDs while Slave signals completion after its 29 local LEDs.
- Correctly map local iteration indices to global state buffers on the Slave half.
- Use 64-bit precision for animation math and maintain volatile state safety.
- Slow down decay to 1 unit per 30ms for a smoother transition.

+++ keyboards/sofle_choc/keymaps/vial/keymap.c
@@ -165,7 +165,7 @@ void matrix_scan_user(void) {
         decay_timer += ticks * 30;
         for (uint8_t i = 0; i < RGB_MATRIX_LED_COUNT; i++) {
             if (led_boost[i] > 127) {
-                uint32_t diff = (uint32_t)(led_boost[i] - 127);
+                uint32_t diff = led_boost[i] - 127;
                 if (diff > ticks) {
                     led_boost[i] -= (uint8_t)ticks;
                 } else {
